+++ lib/rabl-rails/railtie.rb
@@ -3,6 +3,7 @@ module RablRails
     initializer "rabl.initialize" do |app|
       ActiveSupport.on_load(:action_view) do
         ActionView::Template.register_template_handler :rabl, RablRails::Handlers::Rabl
+        RablRails.set_conversion_engines!
       end
     end
   end