+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.2 (unreleased)
   * Add RablRails#render method (see README or source code)
+  * Fix fail when JSON engine is not found. Now fallback to MultiJson.default_adapter
 
 ## 0.1.1